Code cleanup to remove warnings.
Added support for copied IPP string values (save on strdup and
free time)


git-svn-id: svn+ssh://src.apple.com/svn/cups/cups.org/trunk@1409 7a7537e8-13f0-0310-91df-b6672ffda945

+++ pdftops/pdftops.cxx
@@ -88,8 +88,8 @@ int main(int argc, char *argv[]) {
 
     if ((size = ppdPageSize(ppd, NULL)) != NULL)
     {
-      paperWidth  = size->width;
-      paperHeight = size->length;
+      paperWidth  = (int)size->width;
+      paperHeight = (int)size->length;
     }
 
     psOutLevel1 = ppd->language_level == 1;
